fix(Parser/Lexer): fix bug with spans representing empty strs
An empty str is currently represented as a span where begin is equal to (end+1), e.g. (begin: 1, end: 0).
This design might be problematic in the future and we should rethink about other options.

+++ stevia_parser/src/lexer/repr.rs
@@ -41,7 +41,7 @@ impl Span {
     }
 
     pub fn len(self) -> usize {
-        (self.end.to_usize()) - (self.begin.to_usize()) + 1
+        ((self.end.to_usize() as isize) - (self.begin.to_usize() as isize) + 1) as usize
     }
 }
